requirements.txt
Django==1.8.6
gunicorn==19.0.0
mysqlclient==1.3.7
dj-database-url==0.3.0
psycopg2==2.6
dj-static==0.0.6
static3==0.5.1
boto==2.38.0
django-storages==1.1.8
djangorestframework==3.3.1
boto==2.38.0